moving mortgage calculator function tests to unit tests
